refactoring with failing tests
